Feat: 10ft display
I am planning to use this for my living room PC and when I tested on my TV the image took up almost half the screen making it so the accordian and links were not visable. Since I am trying to make this a 10ft display startpage I removed the image to give all of the screen space to the links.